Use non-animiated PNG.
I think the icon was borrowed from the favicon on the Solidus site. The
`ico` format can actually store several images at different resolutions.
When automatically converting it turned these different images into different
frames on an animated PNG. Not all browsers support it but if you view the
icon in Firefox you will see it.

This fixes this by just grabbing the 32x32 image and resizing that down to
24x24 to follow the same size as the previous Spree image.