Reduce setup work for integrating the engines.
The previous instructions recommended the following be added to allow
alchemy content to be available in Solidus views and allow alchemy to
be renderable in Solidus layout:

    # For rendering alchemy content in solidus
    Spree::BaseController.class_eval do
      include Alchemy::ControllerActions
    end

    # Same but only if using solidus-auth-devise
    Spree::UserSessionsController.class_eval do
      include Alchemy::ControllerActionsForSolidus
    end

    # For rendering alchemy pages in Solidus layout
    Alchemy::BaseHelper.send :include, Spree::BaseHelper
    Alchemy::BaseHelper.send :include, Spree::StoreHelper
    Alchemy::BaseController.send :include, Spree::Core::ControllerHelpers::Common
    Alchemy::BaseController.send :include, Spree::Core::ControllerHelpers::Store

To simplify I have made just two requires. It is up to the integrator
if they want either of these or both. If the integrator wants to render
Alchemy content in solidus they add the following to the initializer:

    require 'alchemy/solidus/alchemy_in_solidus'

If you wish to force Alchemy pages to render in the Solidus layout you
add:

    require 'alchemy/solidus/use_solidus_layout'

Of course if you want both you can use both. You no longer need to worry
about solidus-auth-devise. That is now detected and automatically handled.