Update User Extensions to be Reliable and Automatic
Previously a developer needed to manually load the right
decorator depending on if the Alchemy user model or Solidus
user model was in use. Since this configuration can be inspected
refactored those extensions to to be automatic depending on
the config.

This commit also fixes some reloading issues with the extensions.
The documentation was updated to remove the requirement of these
files to be included.

NOTE: This breaks compatability if someone has been requiring those
decorators manually. It should be an easy obvious fix at boot time
but if we really wish to keep compatability we could add two blank
files under the old names.