[ci] exclude alchemy 5.3 and solidus 3.0 from matrix
Alchemy 5.3 is not compatible to Rails 6.1, but Solidus 3.0
needs Rails 6.1 if the default image attachment adapter (Active
Storage) is used. This is the case on CI. Let's skip that build.

If you want to use Alchemy 5.3 and Solidus v3.0, make sure to use
Rails 6.0 and the legacy image attachment adapter (Paperclip)