Add Solidus Alchemy menu items in initializer
This adds the menu items for Alchemy in an `after_initialize` hook
rather than modifying the `spree.rb` initializer. The advantage here is
that we can support both the new MenuItem API from Solidus 4.2+ as well
as the old one. For people who have either not wrapped their in-app
initializer in a `config.after_initialize` hook, we will also add a nice
deprecation warning if they already have the Alchemy menu item in the
list of menu items. For people who have wrapped their backend
configuration in `after_initialize`, there is no hook we can hook into
after that. They will see a double menu entry, and figure out to delete
their own.