Remove navigation deep link to promotions
In the current Solidus `main` branch, the
`Spree::Admin::PromotionsController` does not exist anymore if one only
requires `solidus_core` and `solidus_backend`. It's only present if one
requires the full `solidus` suite with the `solidus_legacy_promotions`
gem.

There's also the new `solidus_promotions` gem, and now the path to
configuring the promotion system depends on the gem being used. By the
time we run this code, the promotion system configuration is probably
not loaded. Given that we don't do anything with promotions in this gem,
let's remove the deep link. Customers can still get at the promotions
admin via the navigation in Solidus.