Manually patch Spree Admin unless new Alchemy
Newer versions of Alchemy do not accidentally load the Admin Base
Controller in an initializer, so for new Alchemy versions this will be
handled by Flickwerk. In order versions though, the
`Spree::Admin::BaseController` is loaded to early in the startup process
and thus we need to load the patch manually.