Add InvalidateElementsCacheJob
In larger applications it might be the case that one product, variant
or taxon is assigned to thousands of elements via one of our
three ingredients. The current implementation does not perform
very well. Using a active job that can be backgrounded and has an
optimized algorithm to touch all elements, their parent elements and
pages should improve performance.