Use full url path for product link
Instead of just using the slug and use a hard coded URL
matching, we add a url_path method to the product
which returns the full product path instead of just the
slug.

Add configuration so that this can be changed to
something else, if you have custom product urls in
your store.